+++ CHANGELOG.md
@@ -182,6 +182,7 @@
 - fixed file streaming encryption
 **1.4.17 : 4-19-2026**
 - finished file streaming decryption
+- adjusted file size function math
 
 ------------------------------------------------------------------------